Remove unnecesary `sudo` from Tavis CI config
`sudo: true` can be used in Travis CI config to specify that our
builds should be send to container-based structure instead of
stardard-based structure. This feature is only available for
repos recognized by Travis CI before 2015/01/01, so this is a feature
unavailable for this repo and therefore `sudo` should be removed.

More info:
http://docs.travis-ci.com/user/workers/container-based-infrastructure/#Routing-your-build-to-container-based-infrastructure